requirements.txt
PyQt5==5.15.10
paho-mqtt==1.6.1